revamped README, first draft
Added a section on displaying the Console IP, or port forwarding to the ClusterIP if no external addresses are available, finished folding in the new content, and tweaked the formatting a bit. I still need to edit the defender install script to allow for passing the IP address as well as the port.