Update namespace_image_collection.sh
Added environment variable and cli option processing, usage text, the ability to overwrite existing Collections, and user prompts to confirm the creation of the new Collection or the replacement of the existing one.

Note that the script uses a few bash-isms so it is not bourne shell (/bin/sh) compatible. Replacing the "[[ ]]" tests with "[ ]" and the array with a character delimited string would probably take care of any portability issues.