[Splunk App] Dynamic conf file generation
Propose POC code to be able to dynamically generate *.conf files.
Initial goal to dynamically generate config files removing the need to touch inputs.conf from windows to nix. 
This is accomplished through the use of os.join.path & configparser. 
A more efficient version of this would be to iterate over the configs but as a demonstration it suffices. 
Additionally this would allow for a interactive install script in the future for defined polling times.
Looking further ahead this + a replacement of index=main in the pcc_incidents.xml would allow for different indexes to be used creating less friction for end user.