Update poll_incidents.py
Reverses offset functionality, in lue of range which counts down and was consistently storing value ≈50. 
With this method we take the total passing it into for loop and iterating over the modulus of 50 if equal to 0 appending it to an array.
Reversing the list we start off with the highest moving to the lowest ends with the highest serial number.